rename (absorbing the little previous content) click_test.go to overlord_test.go because it will be more appropriate soon
